requirements_no_generate.txt
py-cord
python-dotenv
requests
pillow
tomlkit
urlextract
gpt4all
numpy
opencv-python
playwright
tqdm